update: table 컴포넌트 구조 개선
